requirements.txt
python-jsonrpc
pymongo
pika
requests
redis
lxml
sklearn
numpy
scipy
python-dateutil
newspaper
watchdog
pandas
tensorflow
pyyaml
pystatsd
